Fix incorrect print format specifier
`uid_t` is not necessarily `int`. Therefore, `%d` may not necessarily
print it properly. To be safer, cast to `long` and use the corresponding
format specifier, `%ld`. This will still fail if `uid_t` is larger than
`long`, or on very large values if `uid_t` is the same size as `long`
but unsigned, but those are much less likely.

+++ atcd/atcd.c
@@ -212,7 +212,7 @@ static bool run_pending_connection_once(struct connection *conn) {
 		pwd = getpwuid(cred.uid);
 	} while (!pwd && errno == EINTR);
 	if (!pwd) {
-		clprintf(CONN_DEBUG, "[server] user denied for no passwd entry: %d", cred.uid);
+		clprintf(CONN_DEBUG, "[server] user denied for no passwd entry: %ld", (long) cred.uid);
 		clputs(conn, "MATC ACCESS");
 		return false;
 	}
